requirements.txt
streamlit
pillow
pytesseract 
roboflow 
gspread 
google-auth
inference-sdk
google-api-python-client
